Further improving code
This version is quite messy, a lot of comments, the velocity, vertical velocity, x_coordinate, and y_ coordinate of the ball are all stored in a vector for each time step, this will allow the data to be plotted in python later. What remains is to tidy up the code: make sure the return types are correct (bool instead of float), start implementing the terrain stuff so i can add cliffs and things to the simulation.